# openapi.yaml
openapi: 3.1.0
info:
  title: Job Tracker API
  version: 1.0.0
  description: |
    The API behind devvjs.dev/job-tracker. It is the only writer of the tracker's data: the UI,
    the "Job Search" Claude project and the seed script all go through it.

    Paths below are relative to the server URL, e.g. `GET /applications` is
    `GET https://devvjs.dev/job-tracker/api/applications`.

    **Authentication.** Every route needs either the agent key
    (`Authorization: Bearer <TRACKER_AGENT_KEY>`) or the owner's session cookie (`jt_session`),
    except `GET /auth/login`, `GET|POST /auth/callback` and `POST /auth/logout`. A valid bearer key
    wins over a cookie; any other bearer value is 401 even when a valid cookie is also sent. Writes
    are recorded as `claude-project` (agent key), `dakota` (session) or `tracker-app` (server and
    system writes such as the seed).

    **Rate limit.** Each agent key may make `AGENT_RATE_LIMIT_PER_MINUTE` requests (default 60)
    in a fixed one-minute window. Over the limit the API answers 429 with `Retry-After` (seconds).
    Session requests are not limited.

    **Formats.** JSON in and out. POST and PATCH bodies need `Content-Type: application/json`
    (415 otherwise; the auth routes are exempt). Dates are `YYYY-MM-DD` and must be real calendar
    dates. Datetimes are ISO 8601 with an offset on input, with a year from 1000 to 9999 (anything
    else is 400 naming the field); the API returns them in UTC with
...

...

...

...
servers:
...
security:
  - agentKey: []
  - sessionCookie: []
...
paths:
...
  /due:
    get:
      tags: [views]
      operationId: getDue
      summary: What is due on or before a date
      description: |
        Records whose status is not `rejected`, `withdrawn`, `closed` or `accepted`, where
        `next_action_due <= date` or `follow_up_date <= date`. Each item is the list record plus
        `due_on` (the earlier of the two dates) and `overdue` (`due_on < date`). Sorted by `due_on`,
        then id.
      parameters:
        - name: date
          in: query
          description: Defaults to today in the tracker's time zone.
          schema: { type: string, format: date }
      responses:
        "200":
          description: The due list.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DueResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/Internal" }
  /summary:
    get:
      tags: [views]
      operationId: getSummary
      summary: Weekly summary counts
      description: |
        Counts over the inclusive window [`from`, `to`]. `to` defaults to today and `from` to
        `to` minus 6 days. `from` after `to` is 400.
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: The summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Summary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  securitySchemes:
    agentKey:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer <TRACKER_AGENT_KEY>`. Writes are recorded as `claude-project`."
    sessionCookie:
      type: apiKey
      in: cookie
      name: jt_session
      description: The owner's signed session cookie, set by the GitHub sign-in. Writes are recorded as `dakota`.
...
  responses:
    BadRequest:
      description: Validation failed. `error.details` lists `{ path, message }` per problem when there are any.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No valid session and no valid agent key.
      headers:
        WWW-Authenticate:
          schema: { type: string, example: Bearer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: The agent key is over its per-minute limit.
      headers:
        Retry-After:
          description: Seconds until the window resets.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Internal:
      description: An unexpected server error. No stack trace is included.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Status:
      type: string
      description: |
        Pipeline order: watching, shortlisted, preparing, applied, screen, interviewing, offer,
        accepted. Terminal at any point: rejected, withdrawn, closed.
      enum: [watching, shortlisted, preparing, applied, screen, interviewing, offer, accepted, rejected, withdrawn, closed]
    WorkArrangement:
      type: string
      enum: [remote, hybrid, onsite]
    PostingStatus:
      type: string
      enum: [live, removed, unverified]
    MoveTiming:
      type: string
      enum: ["yes", "no", unknown]
    CompSource:
      type: string
      enum: [posting, recruiter, estimate]
    Track:
      type: string
      enum: [ai-engineer, senior-frontend, senior-fullstack, devex-platform, other]
    CompanyStage:
      type: string
      enum: [startup, scaleup, public-product, enterprise, consultancy]
    Source:
      type: string
      enum: [linkedin, builtin, company-site, referral, recruiter-inbound, community, other]
    Referral:
      type: string
      enum: [none, asked, submitted]
    Priority:
      type: string
      enum: [high, medium, low]
    ClosedReason:
      type: string
      enum: [rejected, withdrew, posting-removed, no-response, declined-offer, accepted]
...
    EventType:
      type: string
      enum: [discovered, scored, materials-drafted, applied, email, call, interview, take-home, offer, rejected, withdrew, note, status-change]
    Actor:
      type: string
      enum: [dakota, claude-project, tracker-app]
    Relationship:
      type: string
      enum: [recruiter, referral, hiring-manager, interviewer, peer]

    # ---- nested objects ----------------------------------------------------
    Fit:
      type: object
      description: Fit score. Unknown keys are kept.
      additionalProperties: true
      properties:
        skills: { type: integer, minimum: 0, maximum: 100 }
        experience: { type: integer, minimum: 0, maximum: 100 }
        industry: { type: integer, minimum: 0, maximum: 100 }
        growth: { type: integer, minimum: 0, maximum: 100 }
        total: { type: integer, minimum: 0, maximum: 100, description: Weighted total. }
        weights:
          type: object
          additionalProperties: { type: number }
          example: { skills: 0.35, experience: 0.3, industry: 0.15, growth: 0.2 }
        biggest_risk: { type: string }
        gaps: { type: array, items: { type: string }, example: ["GCP/Vertex AI"] }
        scored_at: { type: string, format: date }
    Materials:
      type: object
      description: Links to the materials sent (not file contents). Unknown keys are kept.
      additionalProperties: true
      properties:
        resume: { type: string }
        cover_letter: { type: string }
        answers: { type: string }
        portfolio_links: { type: array, items: { type: string } }
        email_used: { type: string }
        checklist:
          type: array
          items:
            type: object
            additionalProperties: true
            required: [item, done]
            properties:
              item: { type: string }
              done: { type: boolean }

    # ---- applications ------------------------------------------------------
    Application:
      type: object
      description: |
        An application record. Unknown fields stored in `extra` are returned as additional
        top-level properties. List responses (`GET /applications`, `GET /due`) omit `events`,
        `contacts` and `contact_ids`; `GET /applications/{id}`, write responses and 409/412 bodies
        include them. The export includes `events` and `contact_ids` but not `contacts`.
      additionalProperties: true
      required: [id, company, role_title, work_arrangement, status, discovered_at, created_at, updated_at, updated_by]
      properties:
        id: { type: string, example: "gynger--senior-software-engineer-remote--2026-09-15" }
        company: { type: string }
        role_title: { type: string, description: Exactly as posted. }
        posting_url: { type: [string, "null"], format: uri }
        job_id: { type: [string, "null"] }
        posting_status: { oneOf: [{ $ref: "#/components/schemas/PostingStatus" }, { type: "null" }] }
        posting_verified_at: { type: [string, "null"], format: date }
        jd_snapshot_path: { type: [string, "null"] }
        jd_snapshot: { type: [string, "null"], description: Saved job description text. }
        work_arrangement: { $ref: "#/components/schemas/WorkArrangement" }
        location: { type: [string, "null"], example: "Dearborn, MI" }
        detroit_metro: { type: [boolean, "null"] }
        onsite_requirement: { type: [string, "null"], example: "4+ days/week" }
        remote_scope: { type: [string, "null"], example: "US remote" }
        move_timing_ok: { oneOf: [{ $ref: "#/components/schemas/MoveTiming" }, { type: "null" }] }
        comp_min: { type: [integer, "null"], description: "USD per year, base." }
        comp_max: { type: [integer, "null"], description: "USD per year, base." }
        comp_source: { oneOf: [{ $ref: "#/components/schemas/CompSource" }, { type: "null" }] }
        meets_floor:
          type: [boolean, "null"]
          description: Server-computed. `(comp_max ?? comp_min) >= comp_floor`; null when both are null. Never writable.
        equity_bonus_notes: { type: [string, "null"] }
        track: { oneOf: [{ $ref: "#/components/schemas/Track" }, { type: "null" }] }
        company_archetype: { type: [string, "null"] }
        company_stage: { oneOf: [{ $ref: "#/components/schemas/CompanyStage" }, { type: "null" }] }
        industry: { type: [string, "null"] }
        mission_interest: { type: [array, "null"], items: { type: string }, description: Free tags. }
        fit: { oneOf: [{ $ref: "#/components/schemas/Fit" }, { type: "null" }] }
        source: { oneOf: [{ $ref: "#/components/schemas/Source" }, { type: "null" }] }
        source_detail: { type: [string, "null"] }
        connection: { type: [string, "null"] }
        referral: { oneOf: [{ $ref: "#/components/schemas/Referral" }, { type: "null" }] }
        contact_ids: { type: array, items: { type: string }, description: "Linked contact ids, sorted." }
        status: { $ref: "#/components/schemas/Status" }
        priority: { oneOf: [{ $ref: "#/components/schemas/Priority" }, { type: "null" }] }
        next_action: { type: [string, "null"] }
        next_action_due: { type: [string, "null"], format: date }
        follow_up_date: { type: [string, "null"], format: date }
        discovered_at: { type: string, format: date }
        applied_at: { type: [string, "null"], format: date }
        closed_at: { type: [string, "null"], format: date }
        closed_reason: { oneOf: [{ $ref: "#/components/schemas/ClosedReason" }, { type: "null" }] }
        materials: { oneOf: [{ $ref: "#/components/schemas/Materials" }, { type: "null" }] }
        notes: { type: [string, "null"] }
        project_thread_url: { type: [string, "null"], format: uri, description: Link to the Claude project thread. }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time, description: Send this back as `If-Match` on PATCH. }
        updated_by: { $ref: "#/components/schemas/Actor" }
        events: { type: array, items: { $ref: "#/components/schemas/Event" } }
        contacts: { type: array, items: { $ref: "#/components/schemas/Contact" } }
...
    Event:
      type: object
      required: [id, application_id, at, type, note, by, from_status, to_status]
      properties:
        id: { type: string, format: uuid }
        application_id: { type: string }
        at: { type: string, format: date-time }
        type: { $ref: "#/components/schemas/EventType" }
        note: { type: [string, "null"], description: 'For status-change events, `"<from> → <to>"`.' }
        by: { $ref: "#/components/schemas/Actor" }
        from_status: { oneOf: [{ $ref: "#/components/schemas/Status" }, { type: "null" }], description: Set on status-change events. }
        to_status: { oneOf: [{ $ref: "#/components/schemas/Status" }, { type: "null" }], description: Set on status-change events. }
...
    Contact:
      type: object
      description: A contact. Unknown fields stored in `extra` are returned as additional top-level properties.
      additionalProperties: true
      required: [id, name, created_at, updated_at, updated_by]
      properties:
        id: { type: string, example: "jane-doe--ford" }
        name: { type: string }
        company: { type: [string, "null"] }
        role: { type: [string, "null"] }
        relationship: { oneOf: [{ $ref: "#/components/schemas/Relationship" }, { type: "null" }] }
        linkedin_url: { type: [string, "null"], format: uri }
        email: { type: [string, "null"] }
        last_contact_at: { type: [string, "null"], format: date }
        notes: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        updated_by: { $ref: "#/components/schemas/Actor" }
        application_ids: { type: array, items: { type: string }, description: "Linked application ids, sorted." }
...
    DueItem:
      description: A list record plus when it is due.
      allOf:
        - $ref: "#/components/schemas/Application"
        - type: object
          required: [due_on, overdue]
          properties:
            due_on: { type: string, format: date, description: The earlier of next_action_due and follow_up_date. }
            overdue: { type: boolean, description: "`due_on` is before `date`." }
    DueResponse:
      type: object
      required: [date, items]
      properties:
        date: { type: string, format: date }
        items: { type: array, items: { $ref: "#/components/schemas/DueItem" } }
    SourceRow:
      type: object
      required: [source, applied, responded, interviewed, response_rate]
      properties:
        source: { type: string, description: "The source enum value, or `unknown` when null." }
        applied: { type: integer }
        responded: { type: integer }
        interviewed: { type: integer }
        response_rate: { type: number, description: "responded / applied, rounded to 2 decimals." }
    Summary:
      type: object
      required: [from, to, applied, responses, interviews, by_status, by_source]
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        applied: { type: integer, description: Applications with applied_at in the window. }
        responses:
          type: integer
          description: Distinct applications with a status-change in the window to screen, interviewing, offer or rejected.
        interviews: { type: integer, description: Events of type interview in the window (local dates). }
        by_status:
          type: object
          description: Current count per status, all-time. Every status is present.
          additionalProperties: { type: integer }
        by_source:
          type: array
          description: All-time, over applications with applied_at set; sorted by source.
          items: { $ref: "#/components/schemas/SourceRow" }
...
    Error:
      type: object
      description: The error envelope for every non-2xx JSON response.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - precondition_failed
                - unsupported_media_type
                - precondition_required
                - rate_limited
                - internal
                - upstream_error
            message: { type: string }
            details:
              description: For 400, a list of `{ path, message }` (path is dotted, e.g. `fit.total`).
              type: array
              items:
                type: object
                required: [path, message]
                properties:
                  path: { type: string }
                  message: { type: string }
...